[modify] add get_classifier() and extract common model from
